feat: support inserting links via Telescope-orgmode
This feature provides a convenient way to create links to existing
org-files and headlines via telescope.

It leverages the already existing search for headlines.

It relies on an extension of the API of nvim-orgmode, which exposes its
existing feature to insert links. The Telescope feature acts as a
drop-in-replacement to create/select the Link target.